add authorization

+++ backend/routers/posts.py
@@ -1,5 +1,6 @@
 from typing import Annotated
-from fastapi import APIRouter, Request, HTTPException, status, Depends
+from fastapi import APIRouter, HTTPException, status, Depends
+from auth import CurrentUser
 from schemas import PostCreate, PostResponse
 from sqlalchemy import select
 from sqlalchemy.orm import selectinload
@@ -14,9 +15,14 @@ router = APIRouter()
 @router.delete(
     "/{post_id}", status_code=status.HTTP_204_NO_CONTENT
 )
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(
+    post_id: int,
+    current_user: CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(
-        select(models.Post).options(selectinload(models.Post.author)).where(models.Post.id == post_id)
+        select(models.Post)
+        .options(selectinload(models.Post.author))
+        .where(models.Post.id == post_id)
     )
     post = result.scalars().first()
     if not post:
@@ -24,14 +30,28 @@ async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not Found"
         )
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not Authorized to delete this post."
+        )
+        
     await db.delete(post)
     await db.commit()
 
 @router.post(
     "", response_model=PostResponse, status_code=status.HTTP_201_CREATED
 )
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    new_post = models.Post(**post.model_dump())
+async def create_post(
+    post: PostCreate, 
+    current_user: CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)]):
+    
+    new_post = models.Post(
+        title=post.title,
+        content=post.content,
+        user_id=current_user.id
+    )
     db.add(new_post)
     await db.commit()
     await db.refresh(new_post, attribute_names=["author"])
@@ -41,7 +61,12 @@ async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_
 @router.put(
     "/{post_id}", response_model=PostResponse, status_code=status.HTTP_201_CREATED
 )
-async def update_post_full(post_id: int, post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_full(
+    post_id: int,
+    post: PostCreate,
+    current_user: CurrentUser, 
+    db: Annotated[AsyncSession, Depends(get_db)]):
+    
     result = await db.execute(
         select(models.Post).where(models.Post.id == post_id)
     )
@@ -51,19 +76,15 @@ async def update_post_full(post_id: int, post: PostCreate, db: Annotated[AsyncSe
             detail="Post not Found."
         )
     existing_post = result.scalars().first()
-    if existing_post.user_id != post.user_id:
-        result = await db.execute(
-            select(models.User).where(models.User.id == post.user_id)
+    if existing_post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not Authorized to update this post."
         )
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not Found."
-            )
+        
     existing_post.content = post.content
     existing_post.title = post.title
-    existing_post.user_id = post.user_id
+    existing_post.user_id = current_user.id
     
      
     await db.commit()
@@ -74,7 +95,11 @@ async def update_post_full(post_id: int, post: PostCreate, db: Annotated[AsyncSe
 @router.patch(
     "/{post_id}", response_model=PostResponse, status_code=status.HTTP_201_CREATED
 )
-async def update_post_patch(post_id: int, post: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_patch(
+    post_id: int, 
+    post: PostUpdate, 
+    current_user: CurrentUser, 
+    db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(
         select(models.Post).where(models.Post.id == post_id)
     )
@@ -84,10 +109,17 @@ async def update_post_patch(post_id: int, post: PostUpdate, db: Annotated[AsyncS
             detail="Post not Found."
         )
     existing_post = result.scalars().first()
-    update_data = post.model_dump(exclude_unset=True)
-    for field, value in update_data.items():
-        setattr(existing_post, field, value)
-         
+    if existing_post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not Authorized to update this post."
+        )
+        
+    if post.content:
+        existing_post.content = post.content
+    if post.title:
+        existing_post.title = post.title 
+        
     await db.commit()
     await db.refresh(existing_post, attribute_names=["author"])
     return existing_post
